improve rendering
